# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'LBAW NEWS4U Web API'
  description: 'Web Resources Specification (A7) for NEWS4U'
servers:
- url: http://lbaw2382.fe.up.pt
  description: Production server
...
paths:
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form'
      description: 'Provide login form. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'
      responses:
        '200':
          description: 'Ok. Show log-in UI'

    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Processes the login form submission. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:       # <!--- form field name
                  type: string
                password:    # <!--- form field name
                  type: string
              required:
                - email
                - password

      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful authentication. Redirect to user profile.'
                  value: '/users/{id}'
                302Error:
                  description: 'Failed authentication. Redirect to login form.'
                  value: '/login'
...
  /api/posts:
    get:
      operationId: R201
      summary: 'R201: Search Posts API'
      description: 'Searches for posts and returns the results as JSON. Access: PUB.'

      tags:
        - 'M02: Posts'

      parameters:
        - in: query
          name: query
          description: String to use for full-text search
          schema:
            type: string
          required: false
        - in: query
          name: item
          description: Category of the works
          schema:
            type: string
          required: false
        - in: query
          name: loaned
          description: Boolean with the loaned flag value
          schema:
            type: boolean
          required: false
        - in: query
          name: owner
          description: Boolean with the owner flag value
          schema:
            type: boolean
          required: false
        - in: query
          name: classification
          description: Integer corresponding to the work classification
          schema:
            type: integer
          required: false

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    title:
                      type: string
                    obs:
                      type: string
                    date:
                      type: string
                    owner:
                      type: string
                    tag:
                      type: string
                example:
                  - id: 1
                    title: Ronaldo chega aos 1000 golos
                    obs: ujwkfehf...
                    date: 2025
                    owner: Joana Lima
                    tag: Desporto
                  - id: 15
                    title: Mr. Bean
                    obs: trdyjcghk...
                    date: 1995
                    owner: Manuel Teixeira
                    tag: Filmes

    post:
      operationId: R202
      summary: 'R202: Create Post API'
      description: 'Creates a new post. Access: AUTH.'

      tags:
        - 'M02: Posts'

      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'The title of the post.'
                obs:
                  type: string
                  description: 'Observations or additional information about the post.'
                date:
                  type: string
                  format: date
                  pattern: "DD-MM-YYYY"
                  description: 'The date associated with the post.'
                owner:
                  type: string
                  description: 'The owner of the post.'
                tag:
                  type: string
                  description: 'The tag of the post (e.g., Desporto, Filmes...).'

            examples:
              example1:
                value:
                  title: 'New Music Album'
                  obs: 'Amazing music collection.'
                  date: '2023'
                  owner: 'John Doe'
                  type: 'Música'

      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  message:
                    type: string
...